% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuter.R
\docType{package}
\name{minidebuter-package}
\alias{minidebuter}
\alias{minidebuter-package}
\title{Package pour la formation "Débuter en R"
Contient les vignettes (les slides de cours), le jeu de données "fruits", le jeu de données "nutrimenu"  et les tutoriels. https://vguillemot.github.io/debuter/}
\description{
This package contains all the slides and data-sets used during the course DebuteR.
}
\author{
\strong{Maintainer}: Vincent Guillemot \email{vincent.guillemot@pasteur.fr}

Authors:
\itemize{
  \item Amaury Vaysse \email{amaury.vaysse@pasteur.fr}
}

}
\keyword{internal}
