% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R, R/minidebuter.R
\docType{data}
\name{nutrimenu}
\alias{nutrimenu}
\title{Le jeu de données Nutriwi}
\description{
Composition nutritionnelle de 27 plats alimentaires.

Composition nutritionnelle de 27 plats alimentaires.
}
\details{
La table fruits est de classe "tibble" et contient les colonnes suivantes :
\itemize{
\item ID, identifiant du plat,
\item Type, type de plat (entrée, plat ou dessert),
\item Nutriscore, le nutriscore du plats (A, B ou C)
\item Lien, lien vers le plat sur le site Nutriwi,
\item Glucides, en grammes pour 100g,
\item Lipides, en grammes pour 100g,
\item Proteines, en grammes pour 100g,
\item Sucre, en grammes pour 100g,
\item Fibres, en grammes pour 100g,
\item Sel, en grammes pour 100g,
\item AGS, Acides Gras Saturés, en grammes pour 100g,
\item Energie, en kCal pour 100g.
}

La table fruits est de classe "tibble" et contient les colonnes suivantes :
\itemize{
\item ID, identifiant du plat,
\item Type, type de plat (entrée, plat ou dessert),
\item Nutriscore, le nutriscore du plats (A, B ou C)
\item Lien, lien vers le plat sur le site Nutriwi,
\item Glucides, en grammes pour 100g,
\item Lipides, en grammes pour 100g,
\item Proteines, en grammes pour 100g,
\item Sucre, en grammes pour 100g,
\item Fibres, en grammes pour 100g,
\item Sel, en grammes pour 100g,
\item AGS, Acides Gras Saturés, en grammes pour 100g,
\item Energie, en kCal pour 100g.
}
}
\references{
\url{https://www.nutriwi.com/}

\url{https://www.nutriwi.com/}
}
\keyword{data}
