% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuter.R
\docType{package}
\name{minidebuter-package}
\alias{minidebuter}
\alias{minidebuter-package}
\title{Package léger pour la formation "Débuter en R"
#' @keywords internal}
\description{
Le package minidebuter est le package compagnon de la formation "Débuter en R". Ce package ne contient qu'une seule fonction, les données (fruits et nutrimenu), et les tutoriels.
}
\author{
\strong{Maintainer}: Vincent Guillemot \email{vincent.guillemot@pasteur.fr}

Authors:
\itemize{
  \item Amaury Vaysse \email{amaury.vaysse@pasteur.fr}
}

}
